Copy in flight refactored services as stand-ins
The original intent of this endpoint was to verify the matching & allocating engine we were building. As time has gone (and thinking more of our work on the long lived [Implement alternate version of 2PT matching](https://github.com/DEFRA/water-abstraction-system/pull/458) branch) though doing the same thing the check engine and actual engine were completely separate.

This check endpoint should be using the same engine we are building. So, with this change we remove the duplicate elements, migrate to using the services the refactoring has built and merged into `main`, and use 'stand-ins' for those still in-flight.

It'll mean as the real services are built and merged in we'll need to keep coming back and deleting the stand-ins. Plus what we have in `/check` becomes more about making the output match the scenario format. But it'll be more 'true'.
